requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
cx-Oracle==7.1.3
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.9
django-crispy-forms==1.7.2
django-filer==1.5.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mdeditor==0.1.16
django-mptt==0.10.0
django-polymorphic==2.0.3
django-reversion==3.0.3
easy-thumbnails==2.6
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
kiwisolver==1.1.0
Markdown==3.1.1
markdown2==2.3.8
matplotlib==3.0.3
mysqlclient==1.4.2.post1
numpy==1.16.3
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
reversion==0.2
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
Unidecode==1.0.23
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
